fix: 타입들 관심사에 맞게 분리(me, members, trip, expense)
