+++ src/model/me.ts
@@ -1,0 +1,6 @@
+import { Animals } from 'api/types'
+
+export interface MyInfo {
+  nickName: string;
+  profile: Animals;
+}
